# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for managing inventory",
        "title": "Inventory Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/inventory-details-overall": {
            "get": {
                "description": "Get a comprehensive list of inventory and transaction details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory_overall"
                ],
                "summary": "Fetch overall inventory details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InventoryOverall"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/inventory-details-overall-filter": {
            "get": {
                "description": "Retrieve overall inventory records based on query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "inventory_overall"
                ],
                "summary": "Filter overall inventory details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Product Name",
                        "name": "product_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Brand",
                        "name": "brand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Stock In Hand",
                        "name": "stock_in_hand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Unit Price",
                        "name": "unit_price",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Supplier ID",
                        "name": "supplier_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date Added",
                        "name": "date_added",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Last Updated",
                        "name": "last_updated",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Remarks",
                        "name": "remarks",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "transaction_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Type",
                        "name": "transaction_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Quantity",
                        "name": "quantity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Date",
                        "name": "transaction_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Description",
                        "name": "transaction_description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Amount",
                        "name": "transaction_amount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Status",
                        "name": "transaction_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction Notes",
                        "name": "transaction_notes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.InventoryOverall"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.InventoryOverall": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "date_added": {
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "remarks": {
                    "type": "string"
                },
                "stock_in_hand": {
                    "type": "integer"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "transaction_amount": {
                    "type": "number"
                },
                "transaction_date": {
                    "type": "string"
                },
                "transaction_description": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "integer"
                },
                "transaction_notes": {
                    "type": "string"
                },
                "transaction_status": {
                    "type": "string"
                },
                "transaction_type": {
                    "type": "string"
                },
                "unit_price": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}